# spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: freeq
  version: 0.1.0
  summary: AT Protocol-authenticated IRC server with a JSON API for agents.
  description: |
    freeq is an IRC server where identity is an AT Protocol DID (via the
    `ATPROTO-CHALLENGE` SASL mechanism) rather than a nickname, plus a REST
    surface so agents can read and verify conversations without speaking IRC.

    **Errors** — every `/api/v1/*` 4xx and 5xx answers with the `Error`
    schema: `error` (reason phrase, safe to branch on), `status`, `message`
    (prose, not stable), and `documentation`. A `429` carries `Retry-After`
    in seconds; a `401` carries
    `WWW-Authenticate: Bearer resource_metadata=…` (RFC 9728) naming the
    metadata document that explains how to get a token.

    **Pagination** — list endpoints that can grow use one cursor pattern:
    `limit` (bounded, server-capped at 200) plus `before`, the unix-seconds
    `timestamp` of the oldest item on the page you already have. Absent
    `before` means newest first. Paging by timestamp rather than offset means a
    page cannot shift under a concurrent write. An empty array means the end;
    there is no separate "has more" flag to disagree with the data.

    Note the cursor is the `timestamp` field, not `msgid`: msgids are ULIDs and
    sort the same way, but the server compares timestamps, so passing a msgid
    here silently returns nothing.

    **Versioning and deprecation** — the path carries the major version
...

...

...

...

...
  license:
...
servers:
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /mcp:
    post:
      tags: [discovery]
      operationId: postMcp
      summary: Remote MCP endpoint (Streamable HTTP, JSON-RPC 2.0).
      description: |
        Zero-install MCP: point an MCP client at this URL rather than building
        the stdio server from source. Read-only tools over public
        conversations — `freeq_channels`, `freeq_history`, `freeq_search`,
        `freeq_verify`, `freeq_pins` — each of which calls the same REST
        handler documented here and inherits its authorization. Send
        `Authorization: Bearer <API-BEARER>` to read a restricted channel you
        are a member of.

        Answers a POST with a single `application/json` body; it opens no
        server-initiated SSE stream, so `GET /mcp` is 405.
      security: [{}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: A JSON-RPC 2.0 request, or an array of them.
      responses:
        "200":
          description: JSON-RPC response (or array of responses).
          content:
            application/json:
              schema:
                type: object
        "202":
          description: Accepted — the request was entirely notifications.
  /.well-known/mcp:
    get:
      tags: [discovery]
      operationId: getMcpWellKnown
      summary: MCP endpoint discovery.
      security: [{}]
      responses:
        "200":
          description: Endpoint URL and transport.
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A freeq API bearer token. Obtained either from the `API-BEARER`
        NOTICE the server sends after a successful SASL `903`, or from
        `POST /auth/broker/web-token`. Identifies the caller by DID.
    brokerSignature:
      type: apiKey
      in: header
      name: X-Broker-Signature
      description: HMAC over the raw request body, using the broker shared secret.
...